smoother time checks

+++ stopwatch.php
@@ -1,10 +1,8 @@
 <!DOCTYPE html><html><meta http-equiv="refresh" content="30"><meta name="viewport" content="width=device-width, initial-scale=1"><title>Second Counter</title><style>body{font-family:consolas,sans-serif; font-size:50px;margin-top:60px}#time{border:2px solid #d80; border-radius:12px; width: 8em;padding:0.2em;color:#930;}#s{font-size:16px;margin-left:60px}#days{font-size:30px}</style><body><p id="s">
 <?php
+//$file = 1671579350; (The timestamp on my site)
+if(!file_exists('time2.txt')){$timed = time(); file_put_contents('time2.txt', $timed) or exit("<mark>Can't write</mark>");}
 $file = file_get_contents('time2.txt');
-//$file = 1671579350;
-if($file=="")
-{$timed = time(); file_put_contents('time2.txt', $timed);
-$file = file_get_contents('time2.txt');}
 echo (time() - $file);?>
 
 </p><center><p id="time"></p><p id="days"></p></center>
